New page to deploy backend and see active instances. Refactored code to share it with jupyter homepage
