+++ app/app/services/jupyter/deploy/page.tsx
@@ -23,7 +23,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import ResourceSettingSection from "@/components/services/backend/ResourceSettingSection";
-import { ResourceValueOptions } from "../../backend/interface";
+import { ResourceValueOptions } from "../../backend/deploy/interface";
 import { useRouter } from "next/navigation";
 
 export default function JupyterDeployment() {
