+++ components/services/common/Skeleton.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface SkeletonProps {
+  className: string;
+}
+
+const Skeleton = ({ className }: SkeletonProps) => (
+  <div className={`animate-pulse bg-secondary/40 rounded ${className}`}></div>
+);
+
+export default Skeleton;
